Leaves the parsing of the response to HTTParty
HTTParty's response object has a method #parsed_response that returns a
hash with the response properly parsed. Internally, it uses MultiJSON or
MultiXML, depending on the format you've chosen to use in the request.

I believe that delegating this to HTTParty is interesting because the
HTTP request and it's corresponding parsing are all handled by this HTTParty, which already does it well,  and the Picasa gem
can be dedicated to dealing only with Picasa issues.