Use JSON to read from Picasa for faster parsing
Picasa API allows the use of JSON for rading. Since it's faster to parse
JSON, I'm trying to change the GEM so that it will write to Picasa using
XML and will read using JSON.

These changes are just the beginning. Actually they break the gem, since
it's necessary to fix at least each presenter. The parsed JSON has
attribute names that differ from the ones in the parsed XML, which is
why all of the presenter should be revised.

So far, I've fixed the album presenter. All of the others still need to
be changed.

Presenters' tests need to be fixed as well. So the next step is to fix
the album presenter test, before moving on to the other presenters.